+++ src/utils/calendar.ts
@@ -88,10 +88,10 @@ export const fetchLivePlanning = (): { planning: MauriaEventType[], isTomorrow: 
 
   if (livePlanning.length > 0) {
     if (isTomorrow(livePlanning[0].data.end)) {
-      return {planning: livePlanning, isTomorrow: isTomorrowPlanning};
+      return { planning: livePlanning, isTomorrow: isTomorrowPlanning };
     }
     isTomorrowPlanning = false;
-    return {planning: livePlanning, isTomorrow: isTomorrowPlanning};
+    return { planning: livePlanning, isTomorrow: isTomorrowPlanning };
   }
 
   if (livePlanning.length === 0) {
@@ -135,4 +135,30 @@ export const fetchTomorrowLessons = (): MauriaEventType[] => {
         ).slice(-2)}`,
       });
     });
+};
+
+
+export const fetchEvent = (event: AurionEventType): MauriaEventType => {
+  const data = event.title.split("\n");
+
+  const isCurrent = isInInterval(event.start, event.end);
+
+  const startTime = new Date(event.start);
+  const endTime = new Date(event.end);
+
+  return Object.assign({
+    id: event.id,
+    isCurrent,
+    data: event,
+    title: data[2],
+    type: event.className,
+    room: data[0],
+    teacher: data[5],
+    start: `${("0" + startTime.getHours()).slice(-2)}:${(
+      "0" + startTime.getMinutes()
+    ).slice(-2)}`,
+    end: `${("0" + endTime.getHours()).slice(-2)}:${(
+      "0" + endTime.getMinutes()
+    ).slice(-2)}`,
+  });
 };