+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Fichier de changement de version de l'application !
 
 Toutes les versions sont documentées ici, avec les changements apportés entre chaque version.
  
+
+
+ 
 ## [2.7.0] - 2024-01-X
   
 Première mise à jour de l'application depuis la mise en Open Source du projet.
@@ -12,6 +15,7 @@ Première mise à jour de l'application depuis la mise en Open Source du projet.
 ### Added
 
 - Ajout de la mise à jour automatique du planning au lancement de l'application (7sec d'update après lancement)
+- Modal des évenements du planning (avec les détails + suppression des events perso)
  
 ### Changed
   
@@ -19,8 +23,11 @@ Première mise à jour de l'application depuis la mise en Open Source du projet.
  
 ### Fixed
  
-- /
+- Fix de l'affichage du planning (écrasé sur le côté)
  
+
+
+
 ## [<2.7.0] - 2021-2024
  
 - Toutes les anciennes mises à jour sont documentées dans l'application directement (page Support)